interview-test/interview-test/requirements.txt
aiohttp==3.3.2
async-timeout==3.0.0
atomicwrites==1.1.5
attrs==18.1.0
chardet==3.0.4
click==6.7
Flask==1.0.2
idna==2.7
idna-ssl==1.1.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.2.0
multidict==4.3.1
pluggy==0.6.0
py==1.5.4
pytest==3.6.3
six==1.11.0
Werkzeug==0.14.1
yarl==1.2.6
PyJWT==1.7.1


